batch_resizer/requirements.txt
Pillow == 8.4
